+++ lib/src/fields/unified_async_picker_field.dart
@@ -339,11 +339,11 @@ class _UnifiedAsyncPickerFieldState<T>
         placeholder: widget.placeholder ?? dec.placeholder ?? widget.label,
         labelStyle: dec.labelStyle,
         style: dec.fieldStyle,
-        backgroundColor: dec.backgroundColor ?? Colors.black26,
+        backgroundColor: dec.backgroundColor,
         headerBackgroundColor:
-            dec.headerBackgroundColor ?? dec.backgroundColor ?? Colors.black26,
+          dec.headerBackgroundColor ?? dec.backgroundColor,
         borderRadius:
-            dec.borderRadius ?? const BorderRadius.all(Radius.circular(18)),
+            dec.borderRadius,
         borderSide: dec.borderSide,
         height: dec.height,
         rowLabelRatio: dec.rowLabelRatio,
@@ -701,11 +701,11 @@ class _UnifiedAsyncMultiPickerFieldState<T>
         placeholder: widget.placeholder ?? dec.placeholder ?? widget.label,
         labelStyle: dec.labelStyle,
         style: dec.fieldStyle,
-        backgroundColor: dec.backgroundColor ?? Colors.black26,
+        backgroundColor: dec.backgroundColor,
         headerBackgroundColor:
-            dec.headerBackgroundColor ?? dec.backgroundColor ?? Colors.black26,
+          dec.headerBackgroundColor ?? dec.backgroundColor,
         borderRadius:
-            dec.borderRadius ?? const BorderRadius.all(Radius.circular(18)),
+            dec.borderRadius,
         borderSide: dec.borderSide,
         height: dec.height,
         rowLabelRatio: dec.rowLabelRatio,